feat: add reducer dispatch form
